+++ Assets/Scripts/DamageSystem.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void TakeDamage(int Damage);
+    void TakeDamage(float Damage);
 }
 public interface ICanAttack
 {
-    int GiveDamage();
+    float GiveDamage();
 }
